refactor(line): Sections 일급컬렉션 생성
